bags

+++ controllers/contactsControllers.js
@@ -1,4 +1,3 @@
-const contactsService = require("../services/contactsServices.js");
 const {
   createContactSchema,
   updateContactSchema,
@@ -18,6 +17,11 @@ const getAllContacts = async (req, res, next) => {
 const getOneContact = async (req, res, next) => {
   try {
     const { id } = req.params;
+
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      return res.status(400).json({ message: "Invalid contact ID" });
+    }
+
     const result = await Contact.findById(id);
     if (result) {
       res.status(200).json(result);
@@ -32,6 +36,11 @@ const getOneContact = async (req, res, next) => {
 const deleteContact = async (req, res, next) => {
   try {
     const { id } = req.params;
+
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      return res.status(400).json({ message: "Invalid contact ID" });
+    }
+
     const deletedContact = await Contact.findByIdAndDelete(id);
     if (deletedContact) {
       res.status(200).json(deletedContact);
@@ -45,7 +54,18 @@ const deleteContact = async (req, res, next) => {
 
 const createContact = async (req, res, next) => {
   try {
+    const validationResult = createContactSchema.validate(req.body);
+
+    if (validationResult.error) {
+      return res.status(400).json({ message: validationResult.error.message });
+    }
+
     const result = await Contact.create(req.body);
+
+    if (result instanceof Error) {
+      return res.status(400).json({ message: result.message });
+    }
+
     res.status(201).json(result);
   } catch (error) {
     next(error);
@@ -54,32 +74,34 @@ const createContact = async (req, res, next) => {
 
 const updateContact = async (req, res, next) => {
   try {
-    //-----це щоб ви перевірили що я минулого разу все виправив після прийняття домашки😊
-    // const contactId = req.params.id;
-    // const updatedData = req.body;
-    // if (Object.keys(updatedData).length === 0) {
-    //   return res
-    //     .status(400)
-    //     .json({ message: "Body must have at least one field" });
-    // }
-    // const validatedContacts = updateContactSchema.validate(updatedData);
-    // if (validatedContacts.error) {
-    //   return res.status(400).json({ message: validatedContacts.error.message });
-    // }
-    // const updateContact = await Contact.findByIdAndUpdate({
-    //   contactId,
-    //   updatedData,
-    // });
-    // if (!updateContact) {
-    //   res.status(404).json({ message: "Not found" });
-    // }
-    // res.status(200).json(updateContact);
     const { id } = req.params;
-    const result = await Contact.findByIdAndUpdate(id, req.body, { new: true });
-    if (!result) {
-      res.status(404).json({ message: "Not found" });
+
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      return res.status(400).json({ message: "Invalid contact ID" });
     }
-    res.json(result);
+
+    const updatedData = req.body;
+
+    if (Object.keys(updatedData).length === 0) {
+      return res
+        .status(400)
+        .json({ message: "Body must have at least one field" });
+    }
+
+    const validationResult = updateContactSchema.validate(updatedData);
+    if (validationResult.error) {
+      return res.status(400).json({ message: validationResult.error.message });
+    }
+
+    const result = await Contact.findByIdAndUpdate(id, updatedData, {
+      new: true,
+    });
+
+    if (!result) {
+      return res.status(404).json({ message: "Not found" });
+    }
+
+    res.status(200).json(result);
   } catch (error) {
     next(error);
   }
@@ -88,20 +110,29 @@ const updateContact = async (req, res, next) => {
 const updateStatusContact = async (req, res, next) => {
   try {
     const { contactId } = req.params;
+
+    if (!mongoose.Types.ObjectId.isValid(contactId)) {
+      return res.status(400).json({ message: "Invalid contact ID" });
+    }
+
     const { favorite } = req.body;
+
     if (favorite === undefined || typeof favorite !== "boolean") {
       return res
         .status(400)
         .json({ message: "Invalid 'favorite' value. It must be a boolean." });
     }
+
     const updatedContact = await Contact.findByIdAndUpdate(
       contactId,
       { favorite },
       { new: true }
     );
+
     if (!updatedContact) {
       return res.status(404).json({ message: "Contact not found." });
     }
+
     res.status(200).json(updatedContact);
   } catch (error) {
     next(error);
